requirement.txt
aiofiles @ file:///croot/aiofiles_1683773582346/work
aiosqlite @ file:///croot/aiosqlite_1683773899903/work
anyio @ file:///work/ci_py311/anyio_1676823771847/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///work/ci_py311/argon2-cffi-bindings_1676823553406/work
asgiref==3.7.2
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///work/ci_py311/attrs_1676822758931/work
Automat==22.10.0
Babel @ file:///work/ci_py311/babel_1676825020543/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1681493039619/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
boltons @ file:///work/ci_py311/boltons_1677685195580/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///work/ci_py311/cffi_1676822533496/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
comm @ file:///work/ci_py311/comm_1677709131612/work
conda @ file:///croot/conda_1690494963117/work
conda-content-trust @ file:///work/ci_py311/conda-content-trust_1676851327497/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1685032319139/work/src
conda-package-handling @ file:///croot/conda-package-handling_1685024767917/work
conda_package_streaming @ file:///croot/conda-package-streaming_1685019673878/work
constantly==15.1.0
contourpy==1.1.0
cramjam==2.7.0
crispy-bootstrap5==2023.10
cryptography @ file:///croot/cryptography_1686613057838/work
cssselect==1.2.0
cycler==0.11.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Django==4.2.7
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
entrypoints @ file:///work/ci_py311/entrypoints_1676823319002/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///work/ci_py311_2/python-fastjsonschema_1679340124475/work
fastparquet==2023.8.0
filelock==3.12.4
fonttools==4.42.0
fsspec==2023.9.0
geographiclib==2.0
geopy==2.4.0
hyperlink==21.0.0
idna @ file:///work/ci_py311/idna_1676822698822/work
incremental==22.10.0
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipython @ file:///croot/ipython_1691532092695/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itemadapter==0.8.0
itemloaders==1.1.0
jedi @ file:///work/ci_py311_2/jedi_1679336495545/work
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
jmespath==1.0.1
joblib==1.3.2
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema @ file:///croot/jsonschema_1678983421637/work
jupyter-events @ file:///croot/jupyter_events_1684268042853/work
jupyter-server @ file:///work/ci_py311/jupyter_server_1676824947163/work
jupyter-ydoc @ file:///croot/jupyter_ydoc_1683747223142/work
jupyter_client @ file:///work/ci_py311/jupyter_client_1676823491351/work
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
jupyter_server_fileid @ file:///croot/jupyter_server_fileid_1684273577568/work
jupyter_server_ydoc @ file:///croot/jupyter_server_ydoc_1686767404829/work
jupyterlab @ file:///croot/jupyterlab_1686179668131/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab_server @ file:///croot/jupyterlab_server_1680792501448/work
kiwisolver==1.4.4
libmambapy @ file:///croot/mamba-split_1685993156657/work/libmambapy
lxml @ file:///croot/lxml_1679646450407/work
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.7.2
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mistune @ file:///work/ci_py311/mistune_1676823698136/work
mpmath==1.3.0
nbclassic @ file:///croot/nbclassic_1681756170010/work
nbclient @ file:///work/ci_py311/nbclient_1676823713954/work
nbconvert @ file:///work/ci_py311/nbconvert_1676823854564/work
nbformat @ file:///work/ci_py311/nbformat_1676823523130/work
nest-asyncio @ file:///work/ci_py311/nest-asyncio_1676823382924/work
notebook @ file:///croot/notebook_1690984815942/work
notebook_shim @ file:///work/ci_py311/notebook-shim_1676825053954/work
numpy==1.25.2
packaging @ file:///croot/packaging_1678965309396/work
pandas==2.0.3
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parsel==1.8.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==10.0.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
prometheus-client @ file:///work/ci_py311_2/prometheus_client_1679340520168/work
prompt-toolkit @ file:///work/ci_py311/prompt-toolkit_1676824857154/work
Protego==0.3.0
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
psycopg2==2.9.7
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat @ file:///work/ci_py311/pycosat_1676838522308/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyDispatcher==2.0.7
Pygments @ file:///croot/pygments_1684279966437/work
pyOpenSSL @ file:///croot/pyopenssl_1678965284384/work
pyparsing==3.0.9
pyrsistent @ file:///work/ci_py311/pyrsistent_1676823088752/work
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger @ file:///croot/python-json-logger_1683823803357/work
pytz @ file:///work/ci_py311/pytz_1676825004346/work
PyYAML @ file:///work/ci_py311/pyyaml_1676827445582/work
pyzmq @ file:///work/ci_py311/pyzmq_1676918885689/work
queuelib==1.6.2
requests @ file:///croot/requests_1682607517574/work
requests-file==1.5.1
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
scapy==2.5.0
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.11.0
seaborn==0.12.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
service-identity==23.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///work/ci_py311/sniffio_1676823742530/work
soupsieve @ file:///croot/soupsieve_1680518478486/work
sqlparse==0.4.4
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
terminado @ file:///work/ci_py311/terminado_1677696151350/work
threadpoolctl==3.2.0
tinycss2 @ file:///work/ci_py311/tinycss2_1676823757641/work
tldextract==3.5.0
toolz @ file:///work/ci_py311/toolz_1676827522705/work
tornado @ file:///croot/tornado_1690848263220/work
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets @ file:///work/ci_py311/traitlets_1676823305040/work
Twisted==22.10.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
w3lib==2.1.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///work/ci_py311/websocket-client_1676824890004/work
y-py @ file:///croot/y-py_1683555143488/work
ypy-websocket @ file:///croot/ypy-websocket_1684171737040/work
zope.interface==6.0
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
